requirements.txt
# Core parsing
pdfplumber>=0.10.0
python-docx>=1.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# NLP & similarity
sentence-transformers>=2.6.0
spacy>=3.7.0
scikit-learn>=1.4.0
numpy>=1.26.0

# LLM
openai>=1.30.0
anthropic>=0.25.0

# Web verification / scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# CLI & display
typer>=0.12.0
rich>=13.7.0

# Utilities
python-dateutil>=2.9.0
phonenumbers>=8.13.0
